Add AddressSanitizer support and improve CMake configuration
- Introduced an option to enable AddressSanitizer (ASAN) instrumentation in CMake, enhancing debugging capabilities for memory issues.
- Updated CMakeLists.txt to conditionally compile ASAN options based on the selected build environment.
- Enhanced Vulkan and OpenGL renderer files to ensure compatibility with new ASAN settings.
- Refactored various source files to improve type safety and prevent potential issues with unsigned integer comparisons.

These changes aim to improve the development and debugging experience by providing better tools for memory error detection and enhancing code safety across the renderer.

+++ src/renderers/openglrenderer/tr_cmds.c
@@ -130,8 +130,8 @@ static void *R_GetCommandBufferReserved( int bytes, int reservedBytes ) {
 	bytes = PAD(bytes, sizeof(void *));
 
 	// always leave room for the end of list command
-	if ( cmdList->used + bytes + sizeof( int ) + reservedBytes > MAX_RENDER_COMMANDS ) {
-		if ( bytes > MAX_RENDER_COMMANDS - sizeof( int ) ) {
+	if ( cmdList->used + bytes + (int)sizeof( int ) + reservedBytes > MAX_RENDER_COMMANDS ) {
+		if ( bytes > MAX_RENDER_COMMANDS - (int)sizeof( int ) ) {
 			ri.Error( ERR_FATAL, "R_GetCommandBuffer: bad size %i", bytes );
 		}
 		// if we run out of room, just start dropping commands
